+++ src/Exception/InvalidArgumentException.php
@@ -10,6 +10,8 @@ namespace Zend\ServiceManager\Exception;
 use InvalidArgumentException as SplInvalidArgumentException;
 use Zend\ServiceManager\AbstractFactoryInterface;
 use Zend\ServiceManager\Initializer\InitializerInterface;
+use Zend\ServiceManager\Factory\FactoryInterface;
+use Zend\ServiceManager\Factory\DelegatorFactoryInterface;
 
 /**
  * @inheritDoc
@@ -31,6 +33,20 @@ class InvalidArgumentException extends SplInvalidArgumentException implements Ex
     }
 
     /**
+     * @param mixed $factory
+     * @return self
+     */
+    public static function fromInvalidFactory($factory)
+    {
+        return new self(sprintf(
+            'An invalid factory was registered. Expected a valid function name, '
+            . 'class name, a callable or an instance of "%s", but "%s" was received.',
+            FactoryInterface::class,
+            is_object($factory) ? get_class($factory) : gettype($factory)
+        ));
+    }
+
+    /**
      * @param mixed $abstractFactory
      * @return self
      */
@@ -41,6 +57,26 @@ class InvalidArgumentException extends SplInvalidArgumentException implements Ex
             . ' class name resolving to an implementation of "%s", but "%s" was received.',
             AbstractFactoryInterface::class,
             is_object($abstractFactory) ? get_class($abstractFactory) : gettype($abstractFactory)
+        ));
+    }
+
+    public static function fromInvalidDelegatorFactoryClass($name)
+    {
+        return new self(sprintf(
+            'An invalid delegator factory was registered; resolved to class or function "%s" '
+            . 'which does not exist; please provide a valid function name or class name resolving '
+            . 'to an implementation of %s',
+            $name,
+            DelegatorFactoryInterface::class
+        ));
+    }
+
+    public static function fromInvalidDelegatorFactoryInstance($name)
+    {
+        return new self(sprintf(
+            'A non-callable delegator, "%s", was provided; expected a callable or instance of "%s"',
+            is_object($name) ? get_class($name) : gettype($name),
+            DelegatorFactoryInterface::class
         ));
     }
 }