+++ src/ServiceManager.php
@@ -31,6 +31,9 @@ use function is_string;
 use function spl_autoload_register;
 use function spl_object_hash;
 use function trigger_error;
+use Zend\ServiceManager\Factory\FactoryInterface;
+use Zend\ServiceManager\Initializer\InitializerInterface;
+use Zend\ServiceManager\Factory\DelegatorFactoryInterface;
 
 /**
  * Service Manager.
@@ -226,10 +229,6 @@ class ServiceManager implements ServiceLocatorInterface
                 $this->aliases = $config['aliases'] + $this->aliases;
                 $this->mapAliasesToTargets();
             }
-            if (! empty($config['delegators'])) {
-                $this->delegators = array_merge_recursive($this->delegators, $config['delegators']);
-                $this->delegatorCallbackCache = [];
-            }
             if (! empty($config['factories'])) {
                 $this->factories = $config['factories'] + $this->factories;
             }
@@ -238,6 +237,10 @@ class ServiceManager implements ServiceLocatorInterface
             }
             if (! empty($config['shared'])) {
                 $this->shared = $config['shared'] + $this->shared;
+            }
+            if (! empty($config['delegators'])) {
+                $this->delegators = array_merge_recursive($this->delegators, $config['delegators']);
+                $this->delegatorCallbackCache = [];
             }
             if (! empty($config['lazy_services']['class_map'])) {
                 $this->lazyServices['class_map'] = isset($this->lazyServices['class_map'])
@@ -481,10 +484,11 @@ class ServiceManager implements ServiceLocatorInterface
 
         if (is_string($factory) && class_exists($factory)) {
             $factory = new $factory();
-            if (is_callable($factory)) {
+            if ($factory instanceof FactoryInterface) {
                 $this->factories[$name] = $factory;
+                return $factory($this->creationContext, $name, $options);
             }
-            return $factory($this->creationContext, $name, $options);
+            throw InvalidArgumentException::fromInvalidFactory($factory);
         }
 
         if (is_callable($factory)) {
@@ -531,9 +535,9 @@ class ServiceManager implements ServiceLocatorInterface
 
             if (is_string($delegatorFactory) && class_exists($delegatorFactory)) {
                 $delegatorFactory = ($delegatorFactory === Proxy\LazyServiceFactory::class)
-                ? $this->createLazyServiceDelegatorFactory()
-                : new $delegatorFactory();
-                if (is_callable($delegatorFactory)) {
+                    ? $this->createLazyServiceDelegatorFactory()
+                    : new $delegatorFactory();
+                if ($delegatorFactory instanceof DelegatorFactoryInterface) {
                     $this->delegators[$name][$index] = $delegatorFactory;
                     $creationCallback = function () use ($delegatorFactory, $name, $creationCallback, $options) {
                         return $delegatorFactory($this->creationContext, $name, $creationCallback, $options);
@@ -542,9 +546,9 @@ class ServiceManager implements ServiceLocatorInterface
                 }
             }
             if (is_string($delegatorFactory)) {
-                throw ServiceNotCreatedException::fromInvalidClass($delegatorFactory);
+                throw InvalidArgumentException::fromInvalidDelegatorFactoryClass($delegatorFactory);
             }
-            throw ServiceNotCreatedException::fromInvalidInstance($delegatorFactory);
+            throw InvalidArgumentException::fromInvalidDelegatorFactoryInstance($delegatorFactory);
         }
         // cache the callback for later requests
         $this->delegatorCallbackCache[$name] = $creationCallback;
@@ -640,6 +644,9 @@ class ServiceManager implements ServiceLocatorInterface
         foreach ($this->initializers as $idx => $initializer) {
             if (is_string($initializer) && class_exists($initializer)) {
                 $initializer = new $initializer();
+                if (! $initializer instanceof InitializerInterface) {
+                    throw InvalidArgumentException::fromInvalidInitializer($initializer);
+                }
             }
 
             if (is_callable($initializer)) {
@@ -907,7 +914,7 @@ class ServiceManager implements ServiceLocatorInterface
      * configuration present.
      *
      * @return Proxy\LazyServiceFactory
-     * @throws ServiceNotCreatedException when the lazy service class_map
+     * @throws InvalidArgumentException when the lazy service class_map
      *     configuration is missing
      */
     private function createLazyServiceDelegatorFactory()
@@ -919,7 +926,7 @@ class ServiceManager implements ServiceLocatorInterface
         }
 
         if (! isset($this->lazyServices['class_map'])) {
-            throw new ServiceNotCreatedException('Missing "class_map" config key in "lazy_services"');
+            throw new InvalidArgumentException('Missing "class_map" config key in "lazy_services"');
         }
 
         $factoryConfig = new ProxyConfiguration();