+++ test/CommonServiceLocatorBehaviorsTrait.php
@@ -608,7 +608,7 @@ trait CommonServiceLocatorBehaviorsTrait
             ],
         ]);
 
-        self::expectException(ServiceNotCreatedException::class);
+        self::expectException(InvalidArgumentException::class);
         self::expectExceptionMessage($contains);
         $serviceManager->get(stdClass::class);
     }
@@ -865,6 +865,47 @@ trait CommonServiceLocatorBehaviorsTrait
 
         self::expectException(CyclicAliasException::class);
         $sm->setAlias('alias', 'alias');
+    }
+
+    public function testThrowOnFactoriesNotImplementingFactoryInterface()
+    {
+        $sm = $this->createContainer([
+            'factories' => [
+                'wrongFactory' => InvokableObject::class,
+            ],
+        ]);
+        self::expectException(InvalidArgumentException::class);
+        $sm->get('wrongFactory');
+    }
+
+    public function testThrowOnDelegatorsNotImplementingDelegatorFactoryInterface()
+    {
+        $sm = $this->createContainer([
+            'factories' => [
+                'delegator' => SampleFactory::class,
+            ],
+            'delegators' => [
+                'delegator' => [
+                    InvokableObject::class
+                ],
+            ],
+        ]);
+        self::expectException(InvalidArgumentException::class);
+        $sm->get('delegator');
+    }
+
+    public function testThrowOnInitializersNotImplementingInitializerInterface()
+    {
+        $sm = $this->createContainer([
+            'factories' => [
+                'factory' => SampleFactory::class,
+            ],
+            'initializers' => [
+                InvokableObject::class
+            ]
+        ]);
+        self::expectException(InvalidArgumentException::class);
+        $sm->get('factory');
     }
 
     public function testCoverageDepthFirstTaggingOnRecursiveAliasDefinitions()