+++ test/Exception/InvalidArgumentExceptionTest.php
@@ -12,13 +12,13 @@ use stdClass;
 use Zend\ServiceManager\Initializer\InitializerInterface;
 use Zend\ServiceManager\Exception\InvalidArgumentException;
 use Zend\ServiceManager\AbstractFactoryInterface;
+use Zend\ServiceManager\Factory\DelegatorFactoryInterface;
 
 /**
  * @covers \Zend\ServiceManager\Exception\InvalidArgumentException
  */
 class InvalidArgumentExceptionTest extends TestCase
 {
-
     public function testFromInvalidInitializer()
     {
         $exception = InvalidArgumentException::fromInvalidInitializer(new stdClass());
@@ -29,10 +29,6 @@ class InvalidArgumentExceptionTest extends TestCase
             . '", but "stdClass" was received.',
             $exception->getMessage()
         );
-
-        // because the named constructor does not check if classes or functions exist
-        // or the argument is_callable or an instance of InitializerInterface
-        // we are done here
     }
 
     public function testFromInvalidAbstractFactory()
@@ -42,9 +38,35 @@ class InvalidArgumentExceptionTest extends TestCase
         $this->assertSame('An invalid abstract factory was registered. Expected an instance of or a valid'
             . ' class name resolving to an implementation of "'. AbstractFactoryInterface::class
             . '", but "stdClass" was received.', $exception->getMessage());
+    }
 
-        // because the named constructor does not check if classes or functions exist
-        // or the argument is_callable or an instance of InitializerInterface
-        // we are done here
+    public function testFromInvalidDelegatorFactoryClass()
+    {
+        $exception = InvalidArgumentException::fromInvalidDelegatorFactoryClass(stdClass::class);
+        self::assertInstanceOf(InvalidArgumentException::class, $exception);
+        self::assertSame(
+            sprintf(
+                'An invalid delegator factory was registered; resolved to class or function "%s" '
+                . 'which does not exist; please provide a valid function name or class name resolving '
+                . 'to an implementation of %s',
+                stdClass::class,
+                DelegatorFactoryInterface::class
+            ),
+            $exception->getMessage()
+        );
+    }
+
+    public function testFromInvalidDelegatorFactoryInstance()
+    {
+        $exception = InvalidArgumentException::fromInvalidDelegatorFactoryInstance(stdClass::class);
+        self::assertInstanceOf(InvalidArgumentException::class, $exception);
+        self::assertSame(
+            sprintf(
+                'A non-callable delegator, "%s", was provided; expected a callable or instance of "%s"',
+                'string',
+                DelegatorFactoryInterface::class
+            ),
+            $exception->getMessage()
+        );
     }
 }